+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [v0.9.1](https://github.com/2manymws/rc/compare/v0.9.0...v0.9.1) - 2024-01-12
+### Breaking Changes 🛠
+- No request body is used for cache handling by default. by @k1LoW in https://github.com/2manymws/rc/pull/56
+
+## [v0.9.0](https://github.com/2manymws/rc/compare/v0.8.2...v0.9.0) - 2024-01-12
+### Breaking Changes 🛠
+- If cache is not used, no Age header is given by @k1LoW in https://github.com/2manymws/rc/pull/54
+### Fix bug 🐛
+- Keep and use requests unaffected by the next middleware. by @k1LoW in https://github.com/2manymws/rc/pull/53
+
+## [v0.8.2](https://github.com/2manymws/rc/compare/v0.8.1...v0.8.2) - 2024-01-11
+### Fix bug 🐛
+- Fix badkey log by @k1LoW in https://github.com/2manymws/rc/pull/52
+### Other Changes
+- Log more info by @k1LoW in https://github.com/2manymws/rc/pull/50
+
 ## [v0.8.1](https://github.com/2manymws/rc/compare/v0.8.0...v0.8.1) - 2024-01-11
 ### Fix bug 🐛
 - Fix Set-Cookie handling by @k1LoW in https://github.com/2manymws/rc/pull/48